Simplify modules config for now.
This doesn't work with GCC 14 - But cmake is doing the right thing.  The Ubuntu PPA hasn't been updated to 15 yet for 24.04 yet (only for plucky) so I can't easily test to see if the newer GCC does the right thing or not.

However, the logic that was here doesn't seem like the right thing so I've pared it down.